+++ src/api/data-migration/version-1-to-version-2/update-accounts/remove-account-status-fields.ts
@@ -15,9 +15,9 @@ const removeAccountStatusFields = async (connection: Connection) => {
   console.info(`✅ ${loggingMessage}`);
 
   try {
-    const tables = await Promise.all([removeIsVerifiedField(connection), removeIsBlockedField(connection)]);
+    const queries = await Promise.all([removeIsVerifiedField(connection), removeIsBlockedField(connection)]);
 
-    return tables;
+    return queries;
   } catch (err) {
     console.error(`🚨 error ${loggingMessage} %O`, err);